LP805971: Specify UTF8 encoding in password reset templates
As reported by Liam Whalen in Launchpad bug 805971: "The file
/openils/var/template/password-reset/reset-form.tt2 does not display
utf-8 characters.  I modified it to include '<meta
http-equiv="Content-Type" content="text/html; charset=UTF-8" />' inside
the <head> tag, and the fr-CA version of the page now displays
correctly."

I duplicated the problem in Firefox 5 and confirmed that the addition of
the META tag corrects the problem.

Signed-off-by: Dan Scott <dscott@laurentian.ca>