Support for embedded Content Cafe on details page
Added support for two new Apache environment variables
OILS_CONTENT_CAFE_USER and OILS_CONTENT_CAFE_PASS.  If set, a Content
Cafe pane will be loaded (via iframe) directly into a new "Summaries &
More" section on the record details page of the TT-pac, when the
section is expanded.  In addition to the C. Cafe content, we also show
the 520 bib summary in the same pane.

"Summaries & More" is only visible if either there is 520 data or there
is a suitable identifier (isbn or upc) and the content cafe
configuration is in place.

Signed-off-by: Bill Erickson <berick@esilibrary.com>