Auto-retarget local holds on "new" copy checkin
When a new item is checked in local holds will be searched and re-targeted.

The search and retarget stops when a hold is found locally that the item can fill.

Already-captured holds will not be retargeted.

"New" is (currently) defined as "In progress" at checkin

Retarget occurs:
Only if the copy's circ library is the checkin library
Only if the checkin is an actual checkin (not a renewal)
Only if the checkin is capturing holds and transits

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Jason Etheridge <jason@esilibrary.com>