Allow overriding some hold events when permissions apply
Works fine now for HOLD_EXISTS and HOLD_ITEM_CHECKOUT_OUT, at least

But this was also supposed to work with these org settings:
 circ.holds.hold_has_copy_at.alert
 circ.holds.hold_has_copy_at.block
which are supposed to define behavior when the user wants to place a
(copy- or volume-level only?) hold on an item at a branch that's not his
home_ou but his home_ou has an available copy.

Anyway, the open-ils.circ.title_hold.is_possible method is supposed to
return information about whether local copies are available to support
such a function, but it doesn't seem to do so.  I think.  Will revisit.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>