Alert or block holds on available copy at same pickup lib
This improves upon the earlier commit that attempted to deal with what
to do when either of these org_unit settings are true:

     circ.holds.hold_has_copy_at.alert
     circ.holds.hold_has_copy_at.block

The correct use of these settings are that when /any/ type of hold is
attempted at a given pickup lib, if either of those settings are true in
the context of said pickup lib, and there are copies in an available
status at the pickup lib already, inform the user before placing the
hold.  In the alert case, the user can place the hold anyway with
another click.  In the block case, there's an end on't.
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>