Deleted a line the removed single quotes (') from search strings
The multiclass_query function replaces single quotes with spaces (') ( ).
This causes a problem when toggling between Advanced and Basic search.
If a search is conducted with a ' in it, that quote appears in the
basic search box displayed above the results.  However, if the user
then clicks the 'Advanced' button to toggle to the advanced search
the single quote is missing from the search term.  By removing the line
$query =~ /s/'/ /go; from Biblio.pm this problem is resolved.

Signed-off-by: Liam Whalen <lwhalen@evergreen-dev.catalogue.nrcan.gc.ca>
Signed-off-by: Mike Rylander <mrylander@gmail.com>