Vandelay fix for merging without loading new
This fix is derivative of da3a453442050482a419deff91d3dc92fdbf132a
in master, but with new feature code removed for backport to 2.1
and 2.0.  It does the following:

- renames 'vlUploadQueueAutoImport' to 'vlUploadQueueImportNoMatch' to
  match identical change in master and to increase code clarity
- honors this flag when doing a merge.

Previously, if you checked one of the two auto-merge boxes, any records
which did not merge loaded as new records even if the 'Auto-Import
Non-Colliding Records' box was not checked.  You can now do one, the
other, or both together (auto-importing and merging) as needed.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Mike Rylander <mrylander@gmail.com>