robustify (and klunkify) unified volume copy creator
Disable/style the Item Editor while changes in the volume/item pane are being
considered, and require an explicit press of a Ready Item Editor button to
enable the interface.  This reduces the chance of editing widgets in the item
editor from being ripped out from under you (for example, if you try to go
straight from editing a call number to editing an item attribute).

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>