Bypass install_filter mem leak via direct insertion
See https://rt.cpan.org/Ticket/Display.html?id=46691

On the one hand we have a memory leak, on the other we have a bug caused
by the fix to the memory leak.  Bypass this altogether and insert the
filter manually into the template environment using Template's FILTER
configuration argument.  This has the added bonus of simplifying the
i18n filter code.

Since the filter is manually inserted, it's not necessary (and possibly
counter-productive) to USE the filter directly in the template.

Signed-off-by: Bill Erickson <berick@esilibrary.com>