action for marking items claimed never checked out
Menu action in Patron Items Out for marking items as Claimed Never Checked Out.
Note that this action behaves differently than Claimed Returned, in that the
item is actually checked in and the transaction closed, and will not continue
to show up in Items Out for the patron.

If decide later to provide an option for making the two actions consistent,
there is code in the original working branch, collab/phasefx/claimed_never_cko,
which modifies the patron display and the .checked_out methods to look for
open CLAIMSNEVERCHECKEDOUT transactions.  Some cleanup would be needed to avoid
merge conflicts.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>