Force-encode CGI param values to prevent decode errors
CGI.pm decode()'s parameter values, so if it encounters a value that
is already decoded, it will croak and kill the page.  Proactively encode
paramter values so decode() won't fail.

The better solution may be to determine how we get such data, but
until then, let's allow the t-pac to work on funky data.

Signed-off-by: Bill Erickson <berick@esilibrary.com>