Keep the logging output of the EDI translator by default
There is valuable troubleshooting information there that is otherwise hard to
find.  This particularly matters when the output of an EDI
Action/Trigger template (JEDI) doesn't successfully get translated to
EDIFACT for nonobvious reasons.

In production environments, you might instead pipe to logger so that you can
collect output with syslog, rather than in a flat file.  As long as
output goes *somewhere.*

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Jason Etheridge <jason@esilibrary.com>