Assume username, not barcode, in TPAC login
In the absence of a barcode regex OU setting (opac.barcode_regex), the
TPAC was assuming that the entered value was a barcode. However, the
default JSPAC assumption was that the entered value was a username. This
commit makes the TPAC continue to function with the same assumptions as
the default JSPAC in the absence of a barcode regex OU setting.

Signed-off-by: Dan Scott <dscott@laurentian.ca>