TPAC: Strip the "page" GET param from author searches
mkurl() generates a URL that, by default, retains all of the parameters
that the currnt query string includes - useful for retaining library
search scope, etc. However, searches for authors that are launched by
clicking on the author's name in the search results page / record
details were failing in many cases because the "page" parameter came
along with it - and if there weren't five pages of results for page=5,
for example, then no hits would be shown.

This commit removes the page param from the URL. In other places, all of
the params are stripped (see cnbrowse) but as I assume we want to retain
search scope and potentially other parameters, we'll start with the
smallest possible change.

Signed-off-by: Dan Scott <dscott@laurentian.ca>