Fix "Retarget Local Holds" for non-holdable copies
Specificially:

Wrap holdable check in is_true
Add check for current copy status holdable
Add check for current copy location holdable

The latter stores the fetched location for later use.
The one later use I found updated to check if the location is a reference.

The update copy code already checks if location is a reference and deals
with it properly.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>