Disable Google Analytics in T-PAC by default
The option to enable Google Analytics has been turned into a config
setting in parts/config.tt2 for now, with a place to set the GA account
code.

Also, get consistent with foo.enabled instead of a mix of foo.enable /
foo.enabled in config.tt2.

Also, be a bit more careful about whitespace and hide comments in the
HTML source about things that are disabled (such as ChiliFresh and
Google Analytics) until they are enabled.

Signed-off-by: Dan Scott <dscott@laurentian.ca>