Add pgsql 9.0 target for Debian / Ubuntu, and a Fedora 15 target
People are starting to adopt PostgreSQL 9.0 in Evergreen 2.0 as a
stepping stone towards Evergreen 2.1; make it easier for them by
providing a target in the 2.0 Makefile.install.

Also, while we're in here, remove the Fedora 13 reference and replace it
with Fedora 15, as Fedora 13 is no longer supported and Fedora 15 comes
with the power of PostgreSQL 9.0. OH YEAH!

Signed-off-by: Dan Scott <dscott@laurentian.ca>