Support labelFormat for read-only AutoFieldWidget's
Previously only worked with edit widgets (e.g. drop-downs).  This also
takes into account the fact that different instances of AutoFieldWidget
for the same class and value may have different labelFormat's

One user-visible result of this change is the ACQ read-only fund display
in purchase orders.  Funds will now show the code and year for funds in
the copy grid before and after the data is rendered read-only.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>