Remove /default/ from template paths
The original goal of having a /default/ component to the template path
was to support alternate skins.  However, Template Toolkit supports
alternate skins through a more elegant mechanism of providing template
overlays via template path configuration.  In other words, skin files
will be differentiated by the location of their root path and not
because they live in a different skin directory inside the main
template root.

Because of this, the /default/ component in the template path serves
no purpose and generally just gets in the way.

As a part of this, EGWeb.pm no longer inserts the skin name into the
template file path.

Signed-off-by: Bill Erickson <berick@esilibrary.com>