TPac: Multiple holds in staff client place-holds session
Leave the patron_barcode cookie in place after holds placement succeeds
from the embedded catalog so that the next hold request will use the
same patron.  This allows for multiple hold request to take place within
one session.  The negative is that the cookie sticks around even after a
patron session is closed.  This is normally not a problem, because each
new patron sesion will update the cookie.  (And when the cookie is not
updated, staff still have the option to /not/ place a hold for the
pre-selected user).  However, it would be nice to have the staff client
clear the cookie at the right time.  Will look at that too.

Signed-off-by: Bill Erickson <berick@esilibrary.com>