Display warning message and prompt before PostgreSQL removal
Building on the good work in LP# 801624, use the -P "prompt" flag
instead of -yq for apt(itude) to ensure that the user has the ability to
quit the dependency install and reevaluate whether they want to actually
destroy their existing PostgreSQL databases. Note that this is
specifically in the context of the removal of the database packages,
rather than a global change.

Also, turn the inline comment into a displayed description of what is
happening and why, along with a nice big warning.

Signed-off-by: Dan Scott <dscott@laurentian.ca>