Fix some standing penalty issues
Circulations were using the circ matchpoint org unit for blocking.
That could, depending on the matchpoint that matched, result in staff
not being able to see the penalty that is being blocked for.

In addition, blocking is done based on the full path, not ancestors.
But, backend functions only used ancestors. Changed them to use the
full_path variant instead.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>