Use SUBMIT buttons instead of IMAGE buttons
With a bit of CSS, we can replace <input type="image"> buttons with
<input type="submit"> buttons. The payoff is directly translatable
values and better semantics for screen readers, etc, rather than having
to set up a process of generating buttons in your colour & languages of
choice - as well as fewer HTTP requests for images over the network.

Note that we also replace some of the JavaScript-dependent reset /
cancel links with <input type="reset"> buttons.

Signed-off-by: Dan Scott <dscott@laurentian.ca>