LP#848945 - Hide Empty Libs checkbox
The "Show Libraries With Items" button in Holdings Maintenance shows all the
libs with items regardless of the library/depth filters. One problem is that
when the list refreshes (whenever an item/volume is edited, deleted, or added),
staff using this feature have to re-press the button, which can take a while for
some bibs. This branch removes this button and replaces it with a "sticky"
checkbox labelled "Hide Empty Libs".

This allows you to do almost the same thing; you can set the context library to
Consortium and the depth filter to the lowest depth, which would normally show
all libs, but then use the checkbox to prune out the ones that have no children
and no volumes(*). One possible use case that we'd lose is the ability to show a
specific set of libs that don't already have volumes in addition to all the libs
that do.

* - Also, the test for whether to hide an org or not is very simplistic, but
easy to implement. A more aggressive test would hide a library and all its
descendants if none of the descendants have volumes.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>