Suppress Transits
Add two new OU settings:

circ.transit.suppress_hold
circ.transit.suppress_non_hold

If a transit would be generated between the current library and another and
the appropriate setting is set for both end with the same value the transit
will be suppressed. If either end is set to an empty string then transit
suppression will be disabled, even if both ends are set to an empty string.

In addition, transits from other locations will be received when their
destination shares suppression with the current library.

circ.transit.suppress_hold will cause holds to go to the hold shelf instead
of into transit when in effect. Non-hold transits should not check it. This
should be useful when multiple parts of a library share a hold shelf, but
are split on the org tree.

circ.transit.suppress_non_hold will cause items to reshelve instead of
going into transit. Hold transits should not check it. This should be
useful when multiple parts of a library are in the same building and have
no need to transit items before reshelving/cataloging.

In order to suppress *all* transits both settings need to be set.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Dan Wells <dbw2@calvin.edu>