Create git .mailmap file for Evergreen
This is an initial .mailmap file for Evergreen.

The .mailmap file is used by git-shortlog to "coalesce together
commits by the same person".

In the output of git-shortlog, this eliminates duplicate entries
such as the username@UUID entries from the svn to git migration.

Compare "git shortlog -nes" with and without this .mailmap

Signed-off-by: Jeff Godin <jgodin@tadl.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>