Modified spine_labels.js to retain UTF-8 encoding
Dan Wells posted a fix to the DEV list for receipt printing that allowed the
receipt printer to print UTF-8 characters.  I used the ideas in that fix to
modify spine_labels.js.  In spine_labels.js, the url passed to the preview in
preview_xul_192 was encoding the url with the javascript encode() function.
This function does not preserve UTF-8 encodings.  I changed the function to
encodeURIComponent, and UTF-8 characters are now displayed properly by the
spine label printer code.

Signed-off-by: Liam Whalen <lwhalen@evergreen-dev.catalogue.nrcan.gc.ca>
Signed-off-by: Dan Scott <dscott@laurentian.ca>