LP 851915: Remove references to /openils/lib/perl5
To prevent old versions of the OpenSRF and Evergreen Perl modules from
conflicting with the new versions of these modules (which are now
located in the default system Perl package directories), remove
references to them.

Also, O:W:AddedContent contains an old reference to a distro-specific
Perl directory that we can clean up at the same time.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>