Allow choice of placing hold despite age protect
This alters the backend to watch when so much as one copy failed only due
to age protection. If so, a flag is set and returned back up the calling
tree for examination.

In JSPac when the flag is set an alternate confirm message is shown.

In TTPac when the flag is set the failure message is changed and override
is always allowed for the hold in question.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>