Previously invisible copies showed as beige in the staff client, they stopped doing this in 2.0
This is because in 2.0 the copy counts logic was moved from open-ils.storage.biblio.record_entry.copy_count.staff
to an in-database function wrapped in open-ils.search.biblio.record.copy_count.staff.  Both methods returned various
counts including an 'unshadow' count, however the 'unshadow' in the new database function was not completed and it
has only been returning a simple copy count rather than a count of opac visible copies as it should be to
replicate the old behaviour.

Signed-off-by: James Fournie <jfournie@sitka.bclibraries.ca>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>