LP856567: Mark HTML report output as Unicode-encoded
Non-ASCII characters in report ouput that has been generated in HTML
format appears to be corrupted when viewed in a browser. This can be
manually corrected by switching the "Encoding" browser option, but it
would be better to include <meta charset='utf-8'> in the output HTML to
tell the browser directly.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Mike Rylander <mrylander@gmail.com>