LP857460: Get the generic class scheme when adding vols
In a default Evergreen instance, in the "Add Volumes" interface for a
bib record with both LC and Dewey call numbers, the list of call numbers
from which you can select appears to be limited to Dewey, even though
the "Bib Call #" field is populated with an LC call number.

The problem is that here we're grabbing the first entry from the
asset.call_number_classification table, and you might expect that to be
the "Generic" entry - however, it turns out to be the "Dewey" entry
because the entries are ordered alphabetically by .name(), not by
database ID. (You can test this be changing the name of "Dewey" to
"Vicious" and see that the "Generic" entry is chosen instead).

This commit changes the process to first search the acnc array for a
name() = "Generic"; and then if we don't find that, fall back to the
first entry in the classification scheme.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Dan Wells <dbw2@calvin.edu>