Username Change Limits
Default is "Only allowed to change username if it looks like a barcode".

There is a "Lock Username Changes" option to disallow username changing via
the OPAC entirely, useful if syncing usernames to an outside source.

There is also a username change limit disabling option that, if enabled and
the Lock option is disabled, allows the previous behaviour of unlimited
username changes.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Bill Erickson <berick@esilibrary.com>