Replace self-closing divs with open/close divs
XML != HTML

Traditionally, Vandelay templates were parsed as XML on the server for
DTD replacement.  With the advent of locale::maketext for templates,
Vandelay no longer needs to be parsed as XML.  However, not parsing as
XML leaves any self-closing divs (<div/>) intact in the final source,
which are treated as opening divs in text/html.  This patch replaces
self-closing divs with open/close pairs.

https://bugs.launchpad.net/evergreen/+bug/860676

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>