Fix a problem in the clear holds shelf list printer
This is a "special" interface that you get to via Admin -> For
Developers -> Local Administration in the staff client.

The addition features to one "side" of this hackish interface of label affix
and parts had caused breakage on the other "side."  This fixes that.

Also improve document printing, making it less likely the progress
dialog itself gets printed.

One day this will be replaced with something better.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>