TPac: Facets on results page
Show search facets along the left side of the results page.  Clicking on
a facet re-runs the current search with the  new facet appeneded.
Selected facets are styled to indicate they are selected.  Clicking on a
selected facet removes that facet from the set of selected facets (i.e.
it backs out the facet).

TODO: update CSS to match default Evergreen skin

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>